Bazel: reorganize LFS files and add licensing information
