+++ .gitattributes
@@ -80,7 +80,7 @@ csharp/paket.lock linguist-generated=true
 csharp/.paket/Paket.Restore.targets linguist-generated=true eol=crlf
 
 # ripunzip tool
-/misc/bazel/internal/bin/*/ripunzip* filter=lfs diff=lfs merge=lfs -text
+/misc/bazel/internal/ripunzip/ripunzip-* filter=lfs diff=lfs merge=lfs -text
 
 # swift prebuilt resources
 /swift/third_party/resource-dir/*.zip filter=lfs diff=lfs merge=lfs -text
