+++ misc/bazel/internal/ripunzip/README.md
@@ -1,0 +1,2 @@
+These LFS files are distributions of [ripunzip](https://github.com/google/ripunzip), compiled locally with
+`cargo install ripunzip`. A [copy](./LICENSE.txt) of the ripunzip license is included.
