+++ misc/bazel/internal/zipmerge/zipmerge_main.cpp
@@ -1,4 +1,4 @@
-#include "misc/bazel/internal/bin/zipmerge/zipmerge.h"
+#include "misc/bazel/internal/zipmerge/zipmerge.h"
 
 int main(int argc, const char** argv) {
   return zipmerge_main(argc, argv);
